Fixed Mothur input, made string concatenation cleaner
The string concatenation issue was fairly easy to clean up using String.format()

Getting Mothur to work was another story. It was a lot of trial and error and I am not sure how it was working from my terminal with manual input in the first place but the format for executing Mothur batchfiles is now properly implemented. I may want to try Mothur’s command-line interface but I believe the flexibility offered by a Mothur batchfile will be the greatest.

What needs to be done to the existing mothur function:
Delete the batch file, delete the logs, rename the files that are put in cake files.

I also need to make the name of CakeFiles variable and not static