Finished last batch script for final processing step.
Added dynamic folder names, now the name is based off of the name of the file being processed. There is no more /CakeFiles directory.

Added batch script for the final step of processing inside of MothurScripts

Added basic command-line interface for processing two fastq files.

Added EnvironmentInformation class. This stores all the install directories in one place. However it will be expanded in the future to store version numbers and system information in the future.

Added CommandList class. This stores the only specified process currently, but it can be expanded in the future to store other processes.

Changed Constructor for Commands. Now it takes in an arraylist of the two fastq files, and an environment object

Changed getMerge in Commands. The fastq files no longer have to be specified in the method arguments and now only need to be stored in the new FastqFiles arraylist inside Commands.