Added Setup class, added SaveEnvironment class, updated Cake class
Added setup class now provides a textbased user interface for specifying important details unique to the user’s local machine. The details are then saved by the saveEnvironment class.

Added the saveEnvironment class. This takes the details specified in the setup class and saves them to a “CakeFile” that is stored in the path of the .jar running it is contained. It also is responsible for loading the file at runtime.

Updated Cake class. Being the main entry point for the textbased program, it has now been updated to utilize the pipeline implementation of the program. It also calls saveEnvironment which loads the environment details and startup.