Major Refactoring, Changed Usage of Commands’ methods
Moved numerous methods from Commands class to classes inside of the commands package.

Created a commands package that is responsible for storing classes that are responsible for commands

Created CommandExec class for executing console commands from the static exec method.

Mothurscirpts is now responsible for actually executing the commands. It is also now housed in the commands package

Created InternalUtilities class to house methods used in the palmer jar.

BBToolsCommands is now responsible for actually executing its commands. It is also now housed in the commands package.

Updated CommandList to utilize refactored setup.