Added save phase feature, added remove logfile feature
Added SaveState class. The class is used to save the phase of a current task. The phase is then saved in the output directory as a CakeSave. The save will be marked as done when the analysis on a fastq pair is finished. The phase is loaded automatically by the same class at the beginning of the commandList process.

Updated commandList class. The class now utilizes the SaveState class to save the phase of a particular analysis process at each step. The class also loads the state when starting, so if the process was interrupted, it can be resumed promptly.

Added ability to remove logfiles. The remove logfile command has been added to the end of each mothurscript. It does not currently have the cleanest implementation, but it is easiest to pass it through a mothur batchfile to mothur’s console.